requirements.txt
Django==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.3
django-environ==0.11.2
flake8==7.1.1
isort==5.13.2
